+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.3 đang phát triển
+
+Lỗi #1: không kiểm tra được đầu số `099` của `gMobile` (vuongxuongminh)
+
 # 1.0.2
 
-Enh: Upgrade tel-num pattern of `yiiviet\validator\TelNumValidator` (vuongxuongminh)
+Tính năng: cập nhật `pattern` của `yiiviet\validator\TelNumValidator` (vuongxuongminh)